refactor(notes): extract SD recorder + transcription queue into dictation_notes (W5)
W5 de-bloat increment 2. Separate recording LOGIC from recording UI, then move the
engine out of ui_notes.c. The recording INDICATOR (s_rec_seconds/widgets, rec_timer_cb,
show/hide_recording_indicator, voice_state_cb) is pure UI and STAYS; the engine
(s_rec_file/path/samples/mutex/note_slot/pipeline_armed/sd_task + wav_write_header,
sd_record_task, ui_notes_{start,stop,write_audio,pipeline_arm,pipeline_cancel}_recording,
transcription_queue_task, start_transcription_queue) MOVES to dictation_notes.c.

Interface (dictation_notes.h): read accessors dictation_recording_active/_slot/
_pipeline_armed + setter _pipeline_clear_armed (the one cross-module write) + the
dictation_sd_record_start task starter. Shared UI-owned statics the engine reads
(s_voice_recording, volatile s_sd_rec_running, s_next_rec_id) + REC_DIR are extern in
ui_notes_internal.h. Engine statics stay static (s_rec_file/path would collide with
ui_camera.c). Public record/transcribe prototypes stay in ui_notes.h — external callers
(voice.c/ui_home.c/ui_voice.c/main.c/voice_ws_proto.c) unchanged.

ui_notes.c: 3785 -> 3214 LOC. Adversarial diff review: behavior-preserving (storage
identity, initial values, volatile, mutex critical sections, lv_async marshaling, the
clear-armed write point all verified byte-for-byte vs HEAD). Build + host(voice_dictation)
+ clang-format green. LIVE-VERIFIED on 192.168.1.90: offline dictation -> moved
sd_record_task wrote 0094.wav (6.0s/96000 samples) -> reconnect -> moved
transcription_queue_task uploaded 192044 bytes + Dragon transcribed it; zero crashes,
uptime monotonic across the full record/transcribe cycle.
